Extract resource-box
