+++ src/app/pages/details/common/resource-box/resource-boxes.js
@@ -1,21 +1,9 @@
 import React from 'react';
-import { useToggle } from '~/helpers/data';
 import RawHTML from '~/components/jsx-helpers/raw-html';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faExternalLinkAlt } from '@fortawesome/free-solid-svg-icons/faExternalLinkAlt';
-import { faLock } from '@fortawesome/free-solid-svg-icons/faLock';
-import { faDownload } from '@fortawesome/free-solid-svg-icons/faDownload';
-import { faShoppingCart } from '@fortawesome/free-solid-svg-icons/faShoppingCart';
-import { faExclamationTriangle } from '@fortawesome/free-solid-svg-icons/faExclamationTriangle';
-import Dialog from '~/components/dialog/dialog';
-import CompCopyRequestForm from './request-form/request-form';
-import CustomizationForm from '../customization-form/customization-form';
-import VideoResourceBox from './video-resource-box';
+import ResourceBox from './resource-box';
 import useDetailsContext from '../../context';
-import useUserContext from '~/contexts/user';
-import { FormattedMessage } from 'react-intl';
-import linkhelper from '~/helpers/link';
-import cn from 'classnames';
 import './resource-box.scss';
 
 function CommonsHubBox() {
@@ -55,258 +43,6 @@ function CommonsHubBox() {
     );
 }
 
-function NewLabel() {
-    return (
-        <div className="new-label-container">
-            <span className="new-label">NEW</span>
-        </div>
-    );
-}
-
-// eslint-disable-next-line complexity
-function Top({ isNew, model }) {
-    const description = model.comingSoon ?
-        `<p>${model.comingSoonText}</p>` :
-        model.description;
-
-    return (
-        <div className="top">
-            {isNew && <NewLabel />}
-            {model.k12 && (
-                <img
-                    className="badge"
-                    src="/dist/images/details/k-12-icon@3x.png"
-                    alt="K12 resource"
-                />
-            )}
-            <div className="top-line">
-                <h3 className={model.k12 ? 'space-for-badge' : ''}>
-                    {model.heading}
-                </h3>
-                {model.creatorFest && (
-                    <img
-                        title="This resource was created by instructors at Creator Fest"
-                        src="/dist/images/details/cf-badge.svg"
-                    />
-                )}
-            </div>
-            <RawHTML className="description" html={description} />
-        </div>
-    );
-}
-
-function AccessPending() {
-    return (
-        <span className="left">
-            <FontAwesomeIcon icon={faLock} />
-            <span>Access Pending</span>
-        </span>
-    );
-}
-
-function MissingLink() {
-    return (
-        <span className="left missing-link">
-            <FontAwesomeIcon icon={faExclamationTriangle} />
-            <span>MISSING LINK</span>
-        </span>
-    );
-}
-
-function IbooksDialog({ model, isOpen, toggle }) {
-    const bookModel = useDetailsContext();
-
-    if (!('bookModel' in model)) {
-        model.bookModel = bookModel;
-    }
-
-    function done(event) {
-        event.preventDefault();
-        event.stopPropagation();
-        toggle();
-    }
-
-    return (
-        <Dialog
-            isOpen={isOpen}
-            onPutAway={toggle}
-            title="Request a complimentary iBooks title"
-        >
-            <CompCopyRequestForm model={model} done={done} />
-        </Dialog>
-    );
-}
-
-function CustomizationDialog({ isOpen, toggle }) {
-    const bookModel = useDetailsContext();
-
-    function done(event) {
-        event.preventDefault();
-        event.stopPropagation();
-        toggle();
-    }
-
-    return (
-        <Dialog
-            isOpen={isOpen}
-            onPutAway={toggle}
-            title={bookModel.customizationFormHeading}
-        >
-            <CustomizationForm model={bookModel} done={done} />
-        </Dialog>
-    );
-}
-
-// Adapted from get-this-title interceptLinkClicks
-function interceptLinkClicks(event, book, userModel) {
-    const el = linkhelper.validUrlClick(event);
-
-    if (!el) {
-        return;
-    }
-    const trackThis = userModel?.accounts_id && el.dataset.track;
-
-    if (trackThis) {
-        /* eslint-disable camelcase */
-        event.trackingInfo = {
-            book,
-            account_uuid: userModel.uuid,
-            resource_name: el.dataset.track,
-            contact_id: userModel.salesforce_contact_id
-        };
-        /* eslint-enable camelcase */
-    }
-}
-
-function LeftButton({ model }) {
-    const [isOpen, toggle] = useToggle(false);
-    const isCompCopy = model.link.url.endsWith('comp-copy');
-    const isCustomization = model.link.url.endsWith('customized-modules');
-    const icon =
-        {
-            lock: faLock,
-            download: faDownload,
-            'external-link-alt': faExternalLinkAlt
-        }[model.iconType] || faExclamationTriangle;
-    const { userModel } = useUserContext();
-
-    function openDialog(event) {
-        if (isCompCopy || isCustomization) {
-            event.preventDefault();
-            toggle();
-        } else if (model.bookModel) {
-            interceptLinkClicks(event, model.bookModel.id, userModel);
-        }
-    }
-
-    return (
-        <React.Fragment>
-            <a
-                className="left-button"
-                href={model.link.url}
-                data-local={model.iconType === 'lock'}
-                onClick={openDialog}
-                data-track={model.heading}
-                data-analytics-select-content={model.heading}
-                data-content-type={`Book Resource (${model.resourceCategory})`}
-            >
-                <FontAwesomeIcon icon={icon} />
-                <span>{model.link.text}</span>
-            </a>
-            {isCompCopy && <IbooksDialog {...{ model, isOpen, toggle }} />}
-            {isCustomization && <CustomizationDialog {...{ isOpen, toggle }} />}
-        </React.Fragment>
-    );
-}
-
-function useDoneWaitingForModelChange(model) {
-    const [timeIsUp, toggle] = useToggle(false);
-    const timerRef = React.useRef();
-
-    React.useEffect(
-        () => {
-            window.clearTimeout(timerRef.current);
-            toggle(false);
-            timerRef.current = window.setTimeout(
-                () => toggle(true),
-                250
-            );
-        },
-        [model, toggle]
-    );
-
-    return timeIsUp;
-}
-
-// eslint-disable-next-line complexity
-function LeftContent({ model }) {
-    const { userStatus } = useUserContext();
-    const doneWaiting = useDoneWaitingForModelChange(model);
-
-    if (!model.link) {
-        return <AccessPending />;
-    }
-
-    if (!model.link.url) {
-        return doneWaiting && !model.comingSoon ? <MissingLink /> : null;
-    }
-
-    // logged in but not an instructor
-    if (model.iconType === 'lock' && userStatus.isInstructor === false) {
-        return (
-            <div className="left-no-button">
-                <FontAwesomeIcon icon={faLock} />
-                <FormattedMessage
-                    id="resources.available"
-                    defaultMessage="Only available for verified instructors."
-                />
-            </div>
-        );
-    }
-
-    return <LeftButton model={model} />;
-}
-
-function Bottom({ model }) {
-    if (model.comingSoon && model.iconType === 'lock') {
-        return null;
-    }
-    return (
-        <div className="bottom">
-            <LeftContent model={model} />
-            {model.printLink && (
-                <a className="print-link" href={model.printLink}>
-                    <FontAwesomeIcon icon={faShoppingCart} />
-                    <span>Buy print</span>
-                </a>
-            )}
-        </div>
-    );
-}
-
-function ReferenceNumber({ referenceNumber }) {
-    return (
-        referenceNumber !== null && (
-            <div className="reference-number">{referenceNumber}</div>
-        )
-    );
-}
-
-function ResourceBox({ model }) {
-    const classNames = {
-        double: model.double,
-        'coming-soon': model.comingSoon
-    };
-
-    return (
-        <div className={cn('resource-box', classNames)}>
-            <ReferenceNumber referenceNumber={model.videoReferenceNumber} />
-            <Top model={model} isNew={model.isNew} />
-            <Bottom model={model} />
-        </div>
-    );
-}
-
 export default function ResourceBoxes({ models, includeCommonsHub = false }) {
     return (
         <React.Fragment>
@@ -318,23 +54,3 @@ export default function ResourceBoxes({ models, includeCommonsHub = false }) {
     );
 }
 
-
-export function VideoResourceBoxes({
-    models,
-    blogLinkModels,
-    referenceModels
-}) {
-    return (
-        <React.Fragment>
-            {models.map((model) => (
-                <VideoResourceBox {...{ model }} key={model.video_file} />
-            ))}
-            {blogLinkModels?.map((model) => (
-                <ResourceBox model={model} key={model.heading} />
-            ))}
-            {referenceModels.map((model) => (
-                <ResourceBox model={model} key={model.heading} />
-            ))}
-        </React.Fragment>
-    );
-}
