+++ src/app/pages/details/phone-view/instructor-resources-pane/instructor-resources-pane.js
@@ -4,7 +4,8 @@ import {faSignOutAlt} from '@fortawesome/free-solid-svg-icons/faSignOutAlt';
 import {useNavigate} from 'react-router-dom';
 import {resourceBoxModel, useResources} from '../../common/resource-box/resource-box';
 import FeaturedResourcesSection from '../../common/featured-resources/featured-resources.js';
-import ResourceBoxes, {VideoResourceBoxes} from '../../common/resource-box/resource-boxes';
+import ResourceBoxes from '../../common/resource-box/resource-boxes';
+import VideoResourceBoxes from '../../common/resource-box/video-resource-box';
 import useUserContext from '~/contexts/user';
 import useWindowContext, {WindowContextProvider} from '~/contexts/window';
 import './instructor-resources-pane.scss';
